Move fakes generator into the go-protocol pipeline
Since it's a subset of code generation it doesn't need to be its own
pipeline.  In addition, this will simplify upcoming changes for the
input agnostic adapter.
