+++ packages/autorest.go/src/main.ts
@@ -6,7 +6,6 @@
 import { AutoRestExtension, } from '@autorest/extension-base';
 import { transform } from './transform/transform';
 import { protocolGen } from './generator/generator';
-import { fakeGen } from './fake/fake';
 
 require('source-map-support').install();
 
@@ -14,8 +13,7 @@ export async function main() {
   const pluginHost = new AutoRestExtension();
   pluginHost.add('go-transform', transform);
   pluginHost.add('go-protocol', protocolGen);
-  pluginHost.add('go-fake', fakeGen);
-  await pluginHost.Run();
+  await pluginHost.run();
 }
 
 main();
